Move Export functionality to MatricsScreen
