+++ dashboard/src/main/java/com/adauction/group19/controller/MainMenuController.java
@@ -1,7 +1,6 @@
 package com.adauction.group19.controller;
 
 import com.adauction.group19.utils.ThemeManager;
-import com.adauction.group19.view.ExportScreen;
 import com.adauction.group19.view.InputDataScreen;
 import com.adauction.group19.view.ViewMetricsScreen;
 
@@ -51,20 +50,6 @@ public class MainMenuController {
     public void handleViewMetricsButton(ActionEvent actionEvent) {
         if (stage != null) {
             Scene inputDataScene = ViewMetricsScreen.getScene(stage);
-            stage.setScene(inputDataScene);
-        } else {
-            System.out.println("Stage is not set.");
-        }
-    }
-
-    /**
-     * Handles the View Metrics button. Switches scene to the View Metrics screen.
-     * @param actionEvent The action event.
-     */
-    @FXML
-    public void handleExportButton(ActionEvent actionEvent) {
-        if (stage != null) {
-            Scene inputDataScene = ExportScreen.getScene(stage);
             stage.setScene(inputDataScene);
         } else {
             System.out.println("Stage is not set.");
