+++ dashboard/src/main/resources/fxml/MainMenuScreen.fxml
@@ -24,12 +24,6 @@
                         <Label id="btnViewMetricsLabel" text="📈 View metrics"/>
                     </graphic>
                 </Button>
-
-                <Button styleClass="primary-button" id="btnExport" onAction="#handleExportButton">
-                    <graphic>
-                        <Label id="btnExportLabel" text="📤 Export"/>
-                    </graphic>
-                </Button>
             </VBox>
         </VBox>
